requirements.txt
asgiref==3.7.2
coverage==7.3.0
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
gunicorn==21.2.0
iniconfig==2.0.0
packaging==23.1
parameterized==0.9.0
Pillow==10.0.0
pluggy==1.2.0
PyJWT==1.7.1
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
whitenoise==6.5.0
